refactored blog form modal before implementing update
