+++ client/src/components/form/index.js
@@ -1,0 +1,13 @@
+import TitleForm from './TitleForm';
+import SummaryForm from './SummaryForm';
+import TagsForm from './TagsForm';
+import UploadMediaForm from './UploadMediaForm';
+import MediaTextAreaList from './MediaTextAreaList';
+
+export {
+  TitleForm,
+  SummaryForm,
+  TagsForm,
+  UploadMediaForm,
+  MediaTextAreaList,
+};
